+++ src/DIPS.Xamarin.UI/Converters/ValueConverters/TimeConverter.cs
@@ -14,8 +14,6 @@ namespace DIPS.Xamarin.UI.Converters.ValueConverters
     /// </summary>
     public class TimeConverter : IMarkupExtension, IValueConverter
     {
-        private IServiceProvider m_serviceProvider;
-
         /// <summary>
         ///     The converter format that is used to change the format of the time <see cref="TimeConverter" />
         /// </summary>
@@ -42,34 +40,13 @@ namespace DIPS.Xamarin.UI.Converters.ValueConverters
         [ExcludeFromCodeCoverage]
         public object ProvideValue(IServiceProvider serviceProvider)
         {
-            m_serviceProvider = serviceProvider;
             return this;
         }
 
         /// <inheritdoc />
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            var dateTimeInput = DateTime.MinValue;
-            if (value == null) return string.Empty;
-            if (!(value is DateTime) && !(value is TimeSpan))
-                throw new XamlParseException("The input has to be of type DateTime or TimeSpan").WithXmlLineInfo(
-                    m_serviceProvider);
-
-            switch (value)
-            {
-                case TimeSpan timeSpanInput:
-                    dateTimeInput += timeSpanInput;
-                    break;
-                case DateTime dateTimeValue:
-                    dateTimeInput = IgnoreLocalTime ? dateTimeValue : dateTimeValue.ToLocalTime();
-                    break;
-            }
-
-            return Format switch
-            {
-                TimeConverterFormat.Default => ConvertToDefaultFormat(dateTimeInput, culture),
-                _ => string.Empty
-            };
+            return DateTimeFormatter.FormatTime(value, culture, IgnoreLocalTime, Format);
         }
 
         /// <inheritdoc />
@@ -77,24 +54,6 @@ namespace DIPS.Xamarin.UI.Converters.ValueConverters
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
         {
             throw new NotImplementedException();
-        }
-
-        private static string ConvertToDefaultFormat(DateTime dateTimeInput, CultureInfo culture)
-        {
-            var time = dateTimeInput.ToString("HH:mm", culture);
-            if (culture.IsNorwegian())
-            {
-                var hour = dateTimeInput.ToString("HH", culture);
-                var minutes = dateTimeInput.ToString("mm", culture);
-                time = $"{hour}:{minutes}";
-            }
-
-            if (culture.ThreeLetterWindowsLanguageName.Equals("ENU"))
-            {
-                time = dateTimeInput.ToString("hh:mm tt", culture);
-            }
-
-            return time;
         }
     }
 }
